backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7

# notification
channels==4.1.0
channels-redis==4.2.0

Django==5.0.2
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
dlib==19.24.4
drf-yasg==1.21.7
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.7
inflection==0.5.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
